+++ src/Pages/ProjectPage/Schedule/my/MyTime.tsx
@@ -1,13 +1,6 @@
 import styled from 'styled-components';
-import { useContext, useEffect, useState } from 'react';
-import { useUserContext } from '#/Pages/MainPage/MainPage';
-import { ProjectContext } from '#/hooks/context/projectContext';
-import { DateContext } from '#/hooks/context/dateContext';
-import dayjs from 'dayjs';
-import { Http } from '#/constants/backendURL';
 import { ScheduleWeekResponse } from '#/Types/scheduletype';
-import TimeSchedule from '../TimeSchedule';
-import { useParams } from 'react-router-dom';
+import MyTimeSchedule from './MyTimeSchedule';
 
 const CommonText = styled.div`
     color: #000000;
@@ -37,53 +30,11 @@ const HourTextList = styled.div`
 
 const HoursOfDay = [...Array(16).keys()].map((_, index) => index + 9);
 
-const MyTime = () => {
-    // 유저데이터가져옴
-    const { userData } = useUserContext();
-    const memberId = userData?.memberId;
+interface MyTimeProps {
+    scheduleData: ScheduleWeekResponse | null;
+}
 
-    // 프로젝트 ID 가져옴
-    const { projectId } = useParams();
-    console.log(`myTime projectId: ${projectId}`);
-
-    // 달력 선택 날짜 가져옴
-    const date = useContext(DateContext);
-    const condition = dayjs(date?.selectedDate).format('YYYY-MM-DD') ?? '';
-    console.log(`condition : ${condition}`);
-
-    const [scheduleData, setSchdeuleData] =
-        useState<ScheduleWeekResponse | null>(null);
-    // 스케줄 데이터 가져옴
-    useEffect(() => {
-        const accessToken = localStorage.getItem('access_token');
-        const memberId = localStorage.getItem('member_id');
-        const fetchSchedule = async () => {
-            try {
-                const response = await fetch(
-                    `${Http}/v1/projects/${projectId}/members/${memberId}/schedules?condition=${condition}`,
-                    {
-                        method: 'GET',
-                        headers: {
-                            Accept: '*/*',
-                            Authorization: `Bearer ${accessToken}`,
-                        },
-                    }
-                );
-
-                if (!response.ok) {
-                    throw new Error('Failed to fetch pinned projects');
-                }
-
-                const data = await response.json();
-                console.log('내 스케줄', data.data);
-                setSchdeuleData(data.data);
-            } catch (error) {
-                console.error(error);
-            }
-        };
-        fetchSchedule();
-    }, [projectId, memberId, condition]);
-
+const MyTime: React.FC<MyTimeProps> = ({ scheduleData }) => {
     return (
         <TimeTableDiv>
             <div style={{ display: 'flex', flexDirection: 'column' }}>
@@ -92,7 +43,7 @@ const MyTime = () => {
                         <CommonText key={idx}>{hour}</CommonText>
                     ))}
                 </HourTextList>
-                <TimeSchedule scheduleData={scheduleData} />
+                <MyTimeSchedule scheduleData={scheduleData} />
             </div>
         </TimeTableDiv>
     );
